Split pico_ice_sdk into pico_ice_core & pico_ice_usb
