+++ include/ice_fpga.h
@@ -9,7 +9,6 @@ extern "C" {
 #endif
 
 void ice_fpga_init_clock(uint8_t frequency_mhz);
-void ice_fpga_init_uart(uint32_t frequency_hz);
 void ice_fpga_init(void);
 bool ice_fpga_reset(void);
 void ice_fpga_halt(void);
