+++ src/ice_fpga.c
@@ -19,9 +19,6 @@ void ice_fpga_init(void) {
 
     // Output a clock by default.
     ice_fpga_init_clock(48);
-
-    // Enable the UART by default, allowing early init.
-    ice_fpga_init_uart(115200);
 }
 
 void ice_fpga_halt(void) {
@@ -48,22 +45,4 @@ bool ice_fpga_reset(void) {
 
 void ice_fpga_init_clock(uint8_t freq_mhz) {
     clock_gpio_init(ICE_FPGA_CLOCK_PIN, CLOCKS_CLK_GPOUT0_CTRL_AUXSRC_VALUE_CLK_USB, 48 / freq_mhz);
-}
-
-static void ice_fpga_uart_irq_handler(void) {
-    while (uart_is_readable(uart_fpga)) {
-        tud_cdc_n_write_char(1, uart_getc(uart_fpga));
-        tud_cdc_n_write_flush(1);
-    }
-}
-
-void ice_fpga_init_uart(uint32_t baudrate_hz) {
-    uart_init(uart_fpga, baudrate_hz);
-    gpio_set_function(ICE_FPGA_UART_TX_PIN, GPIO_FUNC_UART);
-    gpio_set_function(ICE_FPGA_UART_RX_PIN, GPIO_FUNC_UART);
-
-    // For forwarding UART packets to USB.
-    uart_set_irq_enables(uart_fpga, true, false);
-    irq_set_enabled(ICE_FPGA_UART_IRQ, true);
-    irq_set_exclusive_handler(ICE_FPGA_UART_IRQ, ice_fpga_uart_irq_handler);
 }
